pyproject.toml
[project]
name = "projectalpha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.25",
    "aiofiles>=25.1.0",
    "chonkie>=1.5.2",
    "datasets>=4.4.2",
    "ddgs>=9.10.0",
    "fal-client>=0.12.0",
    "google-genai>=1.59.0",
    "huggingface-hub>=1.3.1",
    "lancedb>=0.26.1",
    "newspaper3k>=0.2.8",
    "openai>=2.15.0",
    "openinference-instrumentation-agno>=0.1.25",
    "openinference-instrumentation-openai>=0.1.41",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opik>=1.9.78",
    "opik-optimizer>=3.0.1",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
    "tqdm>=4.67.1",
    "yfinance>=1.0",
]
